manage auth listing

+++ app/Http/Controllers/ListingController.php
@@ -47,6 +47,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully');
@@ -57,6 +59,10 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+        // logged in user is the owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -78,8 +84,17 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        // logged in user is the owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
